Rename python scripts for question B3
